rename map, added tests
